requirements.txt
Flask
opencv-python
opencv-contrib-python
numpy
tensorflow
tensorflow-hub
pyserial
flask-socketio
matplotlib
torch
gymnasium
stable-baselines3
shimmy